Update some non-intuitive post filenames
The old Jekyll generator used the default URL slug as the filename.
Since all special characters are stripped from the slug, some titles
produced filenames that didn't caccurately represent the intent
(e.g. "c" when using "C#" or "C++").  I already cleaned these up for
the post URLs by updating the slug in the frontmatter, so this just
renames the files to match.